refactor(resource-tables): extract resource tables loading logic
