--- src/pages/resource-tables/use-resource-tables-poll.js
@@ -1,13 +1,0 @@
-import { usePoll } from '../../hooks/use-poll'
-
-const pollQuery = {
-    resource: 'system/tasks/RESOURCE_TABLE',
-    id: ({ taskId }) => taskId,
-}
-
-export const useResourceTablesPoll = () =>
-    usePoll({
-        query: pollQuery,
-        interval: 3000,
-        checkDone: data => data[0].completed,
-    })
